xwayland: refactor selection code
